requirements.txt
# Django stuff
Django==1.10.7
semver==2.6.0
pytz==2017.2
# PIL support
Pillow==2.9.0
python-dateutil==2.6.1

# mysql database
MySQL-python==1.2.5

django-storages==1.6.5



python-resize-image==1.1.10

# CSky models
django-cachemodel==2.1.7
django-basic-models==3.0.4

django-object-actions==0.8.2

python-memcached==1.58

djangorestframework==3.6.2

# Django Allauth
django-allauth==0.32.0
oauthlib==2.0.2
requests==2.18.1
requests-oauthlib==0.4.2
# Permissions framework
rules==0.4

# OAuth
django-oauth-toolkit==1.0.0

# Some extras
django-cors-headers==1.1.0
django-autoslug==1.9.3
wsgiref==0.1.2
sqlparse==0.1.14
netaddr

# Utilities for working with badges
jsonschema==2.6.0
simplejson==3.6.4

# JSON-LD
PyLD==0.7.1
rfc3987==1.3.4
pypng==0.0.18
jsonfield==2.0.2
PyJWT==1.4.0

# markdown support
Markdown==2.6.8
django-markdownify==0.1.0
bleach==2.0.0

# Badgr related utilities
openbadges-bakery==1.0.0b4

celery==4.1.0
django-celery-results==1.0.1
python-json-logger==0.1.2

# SSL Support
cffi==1.2.1
cryptography
enum34==1.0.4
idna==2.5
ipaddress==1.0.14
pyasn1==0.1.9
pycparser==2.14
six==1.9.0

#svg 2 png support
CairoSVG==1.0.22
cssselect==0.9.2
tinycss==0.4

# Backport of Python 3.0 CSV package with Unicode support
backports.csv==1.0.4

# EncryptedCursorPagination dependencies
more-itertools==2.5.0

openbadges==1.0.3
apispec==0.22.0

apispec-djangorestframework==1.0.5
responses==0.8.1
django_mock_queries==1.0.5

# lti consumer
lti==0.9.2

# Optional AWS dependencies
boto==2.48.0
boto3==1.5.6
django-boto==0.3.12
django-ses==0.8.4
